working on gpg package signing

+++ src/build_pkg.py
@@ -1,4 +1,4 @@
-#!/usr/bin/env python
+#!/usr/bin/python
 # -*- coding: utf-8 -*-
 #
 # build_pkg.py
@@ -43,8 +43,10 @@ import re
 import time
 from multiprocessing import Process
 import src.package as pkgclass
+import src.sign_pkgs as sign_pkgs
+import glob
 
-logger = logging.getLogger(__name__)
+logger = logconf.logger
 SRC_DIR = os.path.dirname(__file__) or '.'
 BASE_DIR = os.path.split(os.path.abspath(SRC_DIR))[0]
 DOC_DIR = os.path.join(BASE_DIR, 'docker_files')
@@ -61,6 +63,8 @@ except Exception as err:
 
 
 def remove(src):
+    if src != str(src):
+        return True
     if os.path.isdir(src):
         try:
             shutil.rmtree(src)
@@ -76,6 +80,13 @@ def remove(src):
     else:
         return True
 
+
+def run_docker_clean(pkg=None):
+    try:
+        doc.remove_container(pkg)
+    except Exception:
+        pass
+    return True
 
 def get_pkgver(pkgobj):
     pkg = pkgobj.name
@@ -269,7 +280,13 @@ def handle_hook(first=False, last=False):
             # subprocess.check_call(['tar', '-cf', nxsq + '.tar', nxsq],
             #                           cwd='/opt/antergos-packages/numix-icon-theme')
             try:
-                if 'numix-icon-theme-square' == pack:
+                if 'numix-icon-theme' == pack:
+                    subprocess.call(['git', 'clone', '/var/repo/NX', 'numix-icon-theme'],
+                                    cwd='/opt/antergos-packages/numix-icon-theme')
+                    logger.info('Creating tar archive for %s' % pack)
+                    subprocess.check_call(['tar', '-cf', 'numix-icon-theme.tar', 'numix-icon-theme'],
+                                          cwd='/opt/antergos-packages/numix-icon-theme')
+                elif 'numix-icon-theme-square' == pack:
                     subprocess.call(['git', 'clone', '/var/repo/NXSQ', nxsq],
                                     cwd='/opt/antergos-packages/numix-icon-theme-square')
                     logger.info('Creating tar archive for %s' % pack)
@@ -300,43 +317,57 @@ def handle_hook(first=False, last=False):
             #     db.set('building', '%s not found in database, adding entry..' % package)
             #     db.set('pkg:%s' % package, True)
             #     db.set('pkg:%s:name' % package, package)
-            db.delete('pkg:%s:deps' % pack.name)
-            for dep in depends:
-                db.rpush('pkg:%s:deps' % pack.name, dep)
+            if depends is not None:
+                db.delete('pkg:%s:deps' % pack.name)
+                for dep in depends:
+                    db.rpush('pkg:%s:deps' % pack.name, dep)
             logger.info('Updating pkgver in databse for %s to %s' % (pack.name, version))
             db.set('building', 'Updating pkgver in databse for %s to %s' % (pack.name, version))
             pack.save_to_db('version', version)
 
-        logger.info('All queued packages are in the database, checking deps to determine build order.')
-        db.set('building', 'Determining build order based on pkg dependancies')
-        check = check_deps(packages)
-        if len(check) > 0:
-            for c in check:
-                logger.info('%s depends on a pkg in this build. Moving it to the end of the queue.' % c)
-                db.lrem('queue', 0, c)
-                db.rpush('queue', c)
-        logger.info('Check deps complete. Starting build_pkgs')
-        db.set('building', 'Check deps complete. Starting build container.')
+            logger.info('All queued packages are in the database, checking deps to determine build order.')
+            db.set('building', 'Determining build order based on pkg dependancies')
+            check = check_deps(packages)
+            if len(check) > 0:
+                for c in check:
+                    logger.info('%s depends on a pkg in this build. Moving it to the end of the queue.' % c)
+                    db.lrem('queue', 0, c)
+                    db.rpush('queue', c)
+            logger.info('Check deps complete. Starting build_pkgs')
+            db.set('building', 'Check deps complete. Starting build container.')
+
     logger.info('[FIRST IS SET]: %s' % first)
     logger.info('[LAST IS SET]: %s' % last)
-    build_pkgs(last)
+    if iso_flag == 'False':
+        build_pkgs(last)
 
 
-def update_main_repo(pkg=None, rev_result=None):
+def update_main_repo(pkg=None, rev_result=None, this_log=None):
     if pkg and rev_result:
+        run_docker_clean()
+        repo = 'antergos'
+        repodir = 'main'
         if rev_result == '2':
             rev_result = 'passed'
         elif rev_result == '3':
             rev_result = 'failed'
+        elif rev_result == 'staging':
+            rev_result = None
+            repo = 'antergos-staging'
+            repodir = 'staging'
+        else:
+            logger.error('[UPDATE REPO FAILED]')
+            return
         db.set('idle', 'False')
         result = '/tmp/result'
         if os.path.exists(result):
             shutil.rmtree(result)
         os.mkdir(result, 0o777)
         command = "/makepkg/build.sh"
-        pkgenv = ["_PKGNAME=%s" % pkg, "_RESULT=%s" % rev_result, "_UPDREPO=True"]
+        pkgenv = ["_PKGNAME=%s" % pkg, "_RESULT=%s" % rev_result, "_UPDREPO=True", "_REPO=%s" % repo, "_REPO_DIR=%s" % repodir]
         db.set('building', 'Updating repo database.')
         container = None
+        run_docker_clean("update_repo")
         try:
             container = doc.create_container("antergos/makepkg", command=command,
                                              name="update_repo", environment=pkgenv,
@@ -369,9 +400,13 @@ def update_main_repo(pkg=None, rev_result=None):
                         'ro': False
                     }
             }, privileged=True)
-            this_log = 'repo_update_log'
+            if this_log is None:
+                this_log = 'repo_update_log'
+                upd_repo = False
+            else:
+                upd_repo = True
             cont = db.get('update_repo_container')
-            stream_process = Process(target=publish_build_ouput, args=(cont, this_log))
+            stream_process = Process(target=publish_build_ouput, args=(cont, this_log, upd_repo))
             stream_process.start()
             doc.wait(container)
         except Exception as err:
@@ -419,7 +454,7 @@ def db_filter_and_add(output=None, this_log=None):
     db.set('%s:content' % this_log, pretty.decode('utf-8'))
 
 
-def publish_build_ouput(container=None, this_log=None):
+def publish_build_ouput(container=None, this_log=None, upd_repo=False):
     if not container or not this_log:
         logging.error('Unable to publish build output. (Container is None)')
         return
@@ -448,24 +483,29 @@ def publish_build_ouput(container=None, this_log=None):
             db.publish('build-output', line)
             db.set('build_log_last_line', line)
 
-    db.publish('build-output', 'ENDOFLOG')
+    if upd_repo:
+        db.publish('build-output', 'ENDOFLOG')
     content = '\n '.join(content)
     pretty = highlight(content, BashLexer(), HtmlFormatter(style='monokai', linenos='inline',
                                                            prestyles="background:#272822;color:#fff;",
                                                            encoding='utf-8'))
+    log_exists = db.get('%s:content' % this_log)
+    if log_exists and log_exists != '':
+        pretty = log_exists + pretty
     db.set('%s:content' % this_log, pretty.decode('utf-8'))
 
 
 def build_pkgs(last=False):
     # Create our tmp directories
-    repo = os.path.join("/tmp", "result")
+    result = os.path.join("/tmp", "result")
     cache = os.path.join("/tmp", "pkg_cache")
-    for d in [repo, cache]:
-        if not os.path.exists(d):
-            os.mkdir(d, 0o777)
+    for d in [result, cache]:
+        if os.path.exists(d):
+            shutil.rmtree(d)
+        os.mkdir(d, 0o777)
     pkglist = db.lrange('queue', 0, -1)
     pkglist1 = ['1']
-    in_dir_last = len([name for name in os.listdir(repo)])
+    in_dir_last = len([name for name in os.listdir(result)])
     db.set('pkg_count', in_dir_last)
     for i in range(len(pkglist1)):
         pkg = db.lpop('queue')
@@ -491,10 +531,7 @@ def build_pkgs(last=False):
         pkg_deps = db.lrange('pkg:%s:deps' % pkg, 0, -1)
         pkg_deps_str = ' '.join(pkg_deps)
         logger.info('pkg_deps_str is %s' % pkg_deps_str)
-        try:
-            doc.remove_container(pkg)
-        except Exception:
-            pass
+        run_docker_clean(pkg)
         try:
             container = doc.create_container("antergos/makepkg", command="/makepkg/build.sh " + pkg_deps_str,
                                              name=pkg, volumes=['/var/cache/pacman', '/makepkg', '/repo', '/pkg',
@@ -519,7 +556,7 @@ def build_pkgs(last=False):
                 DOC_DIR:
                     {
                         'bind': '/makepkg',
-                        'ro': True
+                        'ro': False
                     },
                 '/srv/antergos.info/repo/iso/testing/uefi/antergos-staging/':
                     {
@@ -551,7 +588,7 @@ def build_pkgs(last=False):
                         'bind': '/32build',
                         'ro': False
                     },
-                '/tmp/result':
+                result:
                     {
                         'bind': '/result',
                         'ro': False
@@ -560,7 +597,12 @@ def build_pkgs(last=False):
             cont = db.get('container')
             stream_process = Process(target=publish_build_ouput, args=(cont, this_log))
             stream_process.start()
-            doc.wait(container)
+            result = doc.wait(container)
+            if result is not 0:
+                failed = True
+                logger.error('[CONTAINER EXIT CODE] Container %s exited. Return code was %s' % (pkg, result))
+            else:
+                logger.info('[CONTAINER EXIT CODE] Container %s exited. Return code was %s' % (pkg, result))
         except Exception as err:
             logger.error('Start container failed. Error Msg: %s' % err)
             failed = True
@@ -570,12 +612,23 @@ def build_pkgs(last=False):
         # log_stream = stream.split('\n')
         # db_filter_and_add(log_stream, this_log)
 
-        in_dir = len([name for name in os.listdir(repo)])
-        last_count = int(db.get('pkg_count'))
-        logger.info('last count is %s %s' % (last_count, type(last_count)))
-        logger.info('in_dir is %s %s' % (in_dir, type(in_dir)))
-        if in_dir > last_count and not failed:
-            db.incr('pkg_count', (in_dir - last_count))
+        #in_dir = len([name for name in os.listdir(result)])
+        #last_count = int(db.get('pkg_count'))
+        #logger.info('last count is %s %s' % (last_count, type(last_count)))
+        #logger.info('in_dir is %s %s' % (in_dir, type(in_dir)))
+        if not failed:
+            db.publish('build-output', 'Signing package..')
+            pkgs2sign = glob.glob('/srv/antergos.info/repo/iso/testing/uefi/antergos-staging/%s-**.xz' % pkg)
+            try_sign = sign_pkgs.batch_sign(pkgs2sign)
+            if try_sign:
+                db.publish('build-output', 'Updating staging repo database..')
+                update_main_repo(pkg, 'staging', this_log)
+            else:
+                failed = True
+
+        if not failed:
+            db.publish('build-output', 'Build completed successfully!')
+            #db.incr('pkg_count', (in_dir - last_count))
             db.rpush('completed', build_id)
             db.set('%s:result' % this_log, 'completed')
             db.set('%s:review_stat' % this_log, '1')
@@ -604,7 +657,7 @@ def build_pkgs(last=False):
         if last:
             db.set('idle', "True")
             db.set('building', 'Idle')
-            for f in [repo, cache, '/opt/antergos-packages', '/tmp/32bit', '/tmp/32build']:
+            for f in [result, cache, '/opt/antergos-packages', '/tmp/32bit', '/tmp/32build']:
                 remove(f)
         else:
             db.set('building', 'Starting next build in queue...')
@@ -641,13 +694,15 @@ def build_iso():
         else:
             if os.path.exists(flag):
                 os.remove(flag)
+        nm = 'antergos-iso-%s' % arch
         # Initiate communication with docker daemon
+        run_docker_clean(nm)
         try:
             doc = docker.Client(base_url='unix://var/run/docker.sock', timeout=10)
             iso_container = doc.create_container("antergos/mkarchiso",
                                                  volumes=['/antergos-iso/configs/antergos/out', '/var/run/dbus',
                                                           '/start', '/sys/fs/cgroup'], tty=True,
-                                                 name=['antergos-iso-%s' % arch], cpu_shares=512)
+                                                 name=nm, cpu_shares=512)
             db.set('container', iso_container.get('Id'))
         except Exception as err:
             logger.error("Cant connect to Docker daemon. Error msg: %s", err)
@@ -706,6 +761,9 @@ def build_iso():
             failed = True
             db.set('%s:result' % this_log, 'failed')
             db.rpush('failed', build_id)
+        db.set('idle', 'True')
+        db.set('isoFlag', 'False')
+        db.set('isoBuilding', 'False')
 
     try:
         shutil.rmtree('/opt/antergos-packages')
